--- CMakeSettings.json
@@ -1,46 +1,0 @@
-﻿{
-  "version": 3,
-  "cmakeMinimumRequired": {
-    "major": 3,
-    "minor": 31,
-    "patch": 0
-  },
-  "configurePresets": [
-    {
-      "name": "vs2022-x86-Debug",
-      "displayName": "Visual Studio 2022 - x86 Debug",
-      "generator": "Visual Studio 17 2022",
-      "description": "MSBuild x86 Debug using VS2022",
-      "architecture": "x86",
-      "binaryDir": "${sourceDir}/build/${presetName}",
-      "cacheVariables": {
-        "CMAKE_INSTALL_PREFIX": "${sourceDir}/out/install/${presetName}",
-        "CMAKE_EXPORT_COMPILE_COMMANDS": "ON",
-        "CMAKE_CONFIGURATION_TYPES": "Debug",
-        "CMAKE_BUILD_TYPE": "Debug",
-        "ENABLE_TERMINAL": "ON"
-      },
-      "toolset": "host=x64"
-    }
-  ],
-  "buildPresets": [
-    {
-      "name": "vs2022-x86-Debug",
-      "configurePreset": "vs2022-x86-Debug",
-      "description": "Build using Visual Studio 2022 MSBuild (x86 Debug)",
-      "configuration": "Debug",
-      "jobs": 8
-    }
-  ],
-  "testPresets": [
-    {
-      "name": "vs2022-x86-Debug",
-      "configurePreset": "vs2022-x86-Debug",
-      "description": "Run tests using ctest",
-      "configuration": "Debug",
-      "output": {
-        "outputOnFailure": true
-      }
-    }
-  ]
-}
